==> 0_Install_Packages.R <==
# This code only needs to be run once.
install.packages(c("colorspace", "doParallel", "foreach", "knitr", "grid", "magick", "pROC", "spacesXYZ", "tidyverse", "xml2", "yardstick"), repos="https://cloud.r-project.org")

library(colorspace)
library(doParallel)
library(foreach)
library(grid)
library(knitr)
library(magick)
library(pROC)
library(spacesXYZ)
library(tidyverse)
library(xml2)
library(yardstick)

# Show package versions.
sessionInfo()
